LayoutTests/media/media-source/media-source-real-currenttime-progression.html

<!DOCTYPE html>
<html>
<head>
    <title>media-source-real-currenttime-progression</title>
    <script src="mp4-generator.js"></script>
    <script src="../video-test.js"></script>
    <script>
    // Verifies that currentTime advances during playback. Catches regressions
    // where cross-process time reporting stalls or returns stepwise jumps:
    //   - TimeProgressEstimator interpolation (bug 307515)
    //   - currentTime polling frequency (bug 307495 / Paramount+)

    var source;
    var sourceBuffer;
    var time1;
    var time2;

    async function runTest() {
        findMediaElement();

        source = new MediaSource();
        video.src = URL.createObjectURL(source);
        await waitFor(source, 'sourceopen');

        sourceBuffer = source.addSourceBuffer(MP4.VIDEO_TYPE);

        var initData = MP4.initSegment({
            timescale: 1000,
            tracks: [{ id: 1, type: 'video', timescale: 1000 }]
        });
        sourceBuffer.appendBuffer(initData);
        await waitFor(sourceBuffer, 'updateend');

        // 4 seconds of all-sync video so every frame decodes independently.
        var samples = [];
        for (var i = 0; i < 4; i++) {
            samples.push({
                duration: 1000,
                compositionTimeOffset: 0,
                isSync: true,
            });
        }
        var media = MP4.mediaSegment({
            sequenceNumber: 1,
            tracks: [{ id: 1, type: 'video', baseDecodeTime: 0, samples }]
        });
        sourceBuffer.appendBuffer(media);
        await waitFor(sourceBuffer, 'updateend');

        source.endOfStream();

        run('video.currentTime = 0');
        run('video.play()');
        await waitFor(video, 'playing');

        // Wait for currentTime to advance past 0.
        await testExpectedEventually('video.currentTime', 0, '>', 5000);

        time1 = video.currentTime;

        await new Promise(resolve => setTimeout(resolve, 500));

        time2 = video.currentTime;
        testExpected('time2 > time1', true);

        run('video.pause()');
    }

    window.addEventListener('load', event => {
        runTest().then(endTest).catch(failTest);
    });
    </script>
</head>
<body>
    <video></video>
</body>
</html>
